refactor /badge
