+++ src/features/badges/components/BadgeCard.tsx
@@ -2,8 +2,8 @@ import type { FC } from "react";
 import type { FrameColor, ImageType } from "../../../api/badge/types";
 import { formatDateJP } from "../../common/utils";
 import { useMutateBadge } from "../hooks/useMutateBadge";
-import { Badge } from "./badge/Badge";
-import { BadgeSecret } from "./badge/BadgeSecret";
+import { Badge } from "../../common/components/badge/Badge";
+import { BadgeSecret } from "../../common/components/badge/BadgeSecret";
 
 type Props = {
   id: string;
