+++ src/features/common/components/badge/assets/Bow.tsx
