+++ src/features/common/components/badge/assets/Cat.tsx
