+++ src/features/common/components/badge/assets/Gem.tsx
