+++ src/features/common/components/badge/assets/Horse.tsx
