+++ src/features/common/components/badge/assets/Shield.tsx
