+++ src/features/common/components/badge/assets/Sword.tsx
